[StatTable] Work Start

+++ Assets/Project/Scripts/MainGame/Data/ContentData/EquipmentData.cs
@@ -17,7 +17,7 @@ namespace MainGame.Data.ContentData
 			[SerializeField] private Int32 id;
 			[SerializeField] private String name;
 			[SerializeField] private String part;
-			[SerializeField] private Double hp;
+			[SerializeField] private Single hp;
 			[SerializeField] private Single critical;
 			[SerializeField] private Single haste;
 			[SerializeField] private Single hit;
@@ -27,7 +27,7 @@ namespace MainGame.Data.ContentData
 			public Int32 ID => id;
 			public String Name => name;
 			public String Part => part;
-			public Double HP => hp;
+			public Single HP => hp;
 			public Single Critical => critical;
 			public Single Haste => haste;
 			public Single Hit => hit;
